v 0.3.21_201608081613 ・仕様変更 クリッカーの問に対する回答後の百分率で、小数点以下一桁の表示を、 小数点以下四捨五入に変更しました.

+++ app/src/main/java/jp/pmw/test_en_revolution/questionnaire/QuestionResult.java
@@ -15,7 +15,7 @@ public class QuestionResult {
     @SerializedName("answer_count")
     private String answerCount;
     @SerializedName("ratio")
-    private float ratio;
+    private int ratio;
     @SerializedName("percentage")
     private String percentage;
 
@@ -23,7 +23,7 @@ public class QuestionResult {
         this.answerCount = answerCount;
     }
 
-    public void setRatio(float ratio){
+    public void setRatio(int ratio){
         this.ratio = ratio;
     }
